add registrar checks and move new logic to ns_registrar.py
