+++ scanners/dns-scanner/dns_scanner/dns_scanner.py
@@ -12,6 +12,7 @@ from dns.resolver import NXDOMAIN, NoAnswer, NoNameservers, Resolver, Answer
 from dns.exception import Timeout
 
 from dns_scanner.email_scanners import DKIMScanner, DMARCScanner
+from dns_scanner.ns_registrar import check_ns_delegations, get_registrar_context
 
 logger = logging.getLogger(__name__)
 
@@ -29,6 +30,7 @@ class DNSScanResult:
     rcode: str = None
     query_res: dict = None
     ns_delegations: dict = None
+    registrar_context: dict = None
     resolve_chain: list[list[str]] = None
     resolve_ips: [str] = None
     cname_record: str = None
@@ -195,119 +197,6 @@ def get_wildcard_status(domain: str, resolver: Resolver, a_records: Answer):
     return result
 
 
-def probe_nameserver(
-    where: str, qname: str, qtype: str, recursion_desired: bool, timeout: int
-):
-    query = dns.message.make_query(qname, dns.rdatatype.from_text(qtype), use_edns=True)
-    if not recursion_desired:
-        query.flags &= ~dns.flags.RD
-    return dns.query.udp(query, where=where, timeout=timeout)
-
-
-def get_ns_ip(host: str, resolver):
-    ns_ip = None
-    try:
-        ns_a = resolver.resolve(host, "A")
-        if ns_a:
-            ns_ip = ns_a[0].to_text()
-    except (NoAnswer, NXDOMAIN, NoNameservers, Timeout):
-        ns_ip = None
-
-    if ns_ip is None:
-        try:
-            ns_aaaa = resolver.resolve(host, "AAAA")
-            if ns_aaaa:
-                ns_ip = ns_aaaa[0].to_text()
-        except (NoAnswer, NXDOMAIN, NoNameservers, Timeout):
-            ns_ip = None
-
-    return ns_ip
-
-
-def check_ns_delegations(domain, zone_apex, ns_records, resolver=None, timeout_sec=10):
-    if resolver is None:
-        resolver = dns.resolver.get_default_resolver()
-
-    qname = zone_apex
-    if not zone_apex:
-        qname = domain
-
-    ns_hosts = ns_records.get("hostnames", [])
-    if len(ns_records) == 0:
-        try:
-            ns_res = resolver.resolve(domain, dns.rdatatype.NS)
-            ns_hosts = [host.to_text() for host in ns_res]
-        except (NoAnswer, NXDOMAIN, NoNameservers, Timeout):
-            ns_hosts = []
-
-    # Always return structured output, even when empty
-    output = {
-        "ns_hosts": ns_hosts,
-        "ns_checks": [],  # per-NS evidence rows
-        "ns_delegation": {
-            "total_ns": len(ns_hosts),
-            "authoritative_ok": 0,
-            "lame_count": 0,
-            "lame_type": "none",  # none | partial | full
-        },
-    }
-    if len(ns_hosts) == 0:
-        output["ns_delegation"]["lame_type"] = "unknown"
-        return output
-
-    for host in ns_hosts:
-        row = {
-            "ns_host": host,
-            "qname": qname,
-            "qtype": "SOA",
-            "rcode": None,
-            "answered_authoritatively": False,
-            "error": None,
-            "timeout": False,
-        }
-
-        try:
-            ns_ip = get_ns_ip(host, resolver)
-            if ns_ip is None:
-                row["error"] = "ns_ip_resolution_failed"
-                output["ns_delegation"]["lame_count"] += 1
-                output["ns_checks"].append(row)
-                continue
-
-            res = probe_nameserver(ns_ip, domain, "SOA", False, timeout_sec)
-            row["rcode"] = dns.rcode.to_text(res.rcode())
-            row["answered_authoritatively"] = bool(res.flags & dns.flags.AA)
-
-            if row["answered_authoritatively"] and row["rcode"] in [
-                "NOERROR",
-                "NXDOMAIN",
-            ]:
-                output["ns_delegation"]["authoritative_ok"] += 1
-            else:
-                output["ns_delegation"]["lame_count"] += 1
-        except Timeout:
-            row["timeout"] = True
-            row["error"] = "timeout"
-            output["ns_delegation"]["lame_count"] += 1
-        except Exception as e:
-            row["error"] = str(e)
-            output["ns_delegation"]["lame_count"] += 1
-
-        output["ns_checks"].append(row)
-
-    ok = output["ns_delegation"]["authoritative_ok"]
-    total = output["ns_delegation"]["total_ns"]
-
-    if ok == total:
-        output["ns_delegation"]["lame_type"] = "none"
-    elif ok == 0:
-        output["ns_delegation"]["lame_type"] = "full"
-    else:
-        output["ns_delegation"]["lame_type"] = "partial"
-
-    return output
-
-
 def scan_domain(domain, dkim_selectors=None):
     """
     Scan a domain for DNS records
@@ -444,6 +333,12 @@ def scan_domain(domain, dkim_selectors=None):
         domain=domain, zone_apex=zone_apex, ns_records=ns_records
     )
     scan_result.ns_delegations = ns_delegations
+
+    registrar_domain = scan_result.base_domain or zone_apex or domain
+    scan_result.registrar_context = get_registrar_context(
+        base_domain=registrar_domain,
+        ns_hosts=ns_delegations.get("ns_hosts", []),
+    )
 
     # If no MX records are found (with warnings), but there are CNAME records, check the CNAME target for MX records
     if (
